rename globals to helpers refactor usages to new name
